Extract a reusable test setup for normalization test classes
